python/style_transfer_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pillow
numpy
tifffile
torch
diffusers
transformers
accelerate
safetensors
opencv-python
rawpy
protobuf>=3.20.0
# 可选：用于 IP-Adapter
# huggingface-hub
